Only include new interfaces in the advanced interfaces, remove templateBuilder
